Some important changes:
(1) Added back the ability for an agent to make a "do nothing" action.
(2) Brought back the OLD versions of the CC, HPCC and SC clients and made sure they are working as expected.
(3) Created a class, tools.test.EnvironmentTransformer, to transform an environment file from the old version to the new one (using node instead of vertex, for example) and applied it to all current environment files in the project.
(4) Created a class, tools.test.TestConfigurationsGenerator, for generating environment files with random priorities and random agents' positions, to help on future experiments.
(5) Changed the simulator to use floating point values for priorities, instead of integers. Also changed the default value from 0 to 1, as we're going to multiply this value to others.
(6) Changed the files in tools.metrics to consider node priorities and added the quadratic mean metric.
(7) Added back util.net_OLD/ClientConnection_OLD, so it would stop showing annoying error.